Create install guide for installing/running project (#6)
* Create install guide for installing/running project

This commit creates the install guide for installing the required
modules, and running the React project.

* Updated instructions due to change to package.json

This commit changes the instructions to indicate that the image annotation library doesn't need to be "installed" manually anymore. The library does still need to be built but once that occurs, the library can be installed by running npm install in the frontend directory. This is because the custom library has now been added as a dependency of the project in package.json.

* Updated instructions

This commit updates the instructions due to the changes in the frontend's package.json. Now the ufdl-js-client will be installed simply by running npm install in the frontend directory.